+++ common/src/fft_prep_cpx.c
@@ -5,13 +5,13 @@
 #include <stdlib.h>
 
 
-void first_rou(int grade, long double *data){
+void first_rou(int grade, double *data){
   data[0]=cos(SIGN*2*PI/grade);
   data[1]=sin(SIGN*2*PI/grade);
 }
 
-long double* every_rou(int grade){
-    long double *roots=(long double*)calloc(grade, 2*sizeof(long double));
+double* every_rou(int grade){
+    double *roots=(double*)calloc(grade, 2*sizeof(double));
     int it;
     roots[0]=1.; 
     roots[1]=0.;
